requirements.txt
django
djangorestframework
djangorestframework-jwt
django-filter
Pillow
django-cors-headers
openpyxl